+++ packages/core/legos/src/sources.ts
@@ -29,6 +29,7 @@ import * as cfbenchmarks from '@chainlink/cfbenchmarks-adapter'
 import * as cfbenchmarks_test from '@chainlink/cfbenchmarks-test-adapter'
 import * as chain_reserve_wallet from '@chainlink/chain-reserve-wallet-adapter'
 import * as coinapi from '@chainlink/coinapi-adapter'
+import * as coinapi_test from '@chainlink/coinapi-test-adapter'
 import * as coinbase from '@chainlink/coinbase-adapter'
 import * as coincodex from '@chainlink/coincodex-adapter'
 import * as coingecko from '@chainlink/coingecko-adapter'
@@ -100,7 +101,7 @@ import * as metalsapi from '@chainlink/metalsapi-adapter'
 import * as mock_ea from '@chainlink/mock-ea-adapter'
 import * as mycryptoapi from '@chainlink/mycryptoapi-adapter'
 import * as ncfx from '@chainlink/ncfx-adapter'
-import * as ncfx_test from '@chainlink/ncfx-test-adapter'
+import * as nft_blue_chip from '@chainlink/nft-blue-chip-adapter'
 import * as nikkei from '@chainlink/nikkei-adapter'
 import * as nomics from '@chainlink/nomics-adapter'
 import * as oilpriceapi from '@chainlink/oilpriceapi-adapter'
@@ -176,6 +177,7 @@ export default {
   cfbenchmarks_test,
   chain_reserve_wallet,
   coinapi,
+  coinapi_test,
   coinbase,
   coincodex,
   coingecko,
@@ -247,7 +249,7 @@ export default {
   mock_ea,
   mycryptoapi,
   ncfx,
-  ncfx_test,
+  nft_blue_chip,
   nikkei,
   nomics,
   oilpriceapi,