+++ packages/sources/nft-blue-chip/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export { endpoint as marketcapEndpoint } from './marketcap'