+++ packages/composites/augur/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/augur-adapter
 
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/sportsdataio-adapter@1.3.19
+  - @chainlink/therundown-adapter@1.3.27
+
 ## 2.0.10
 
 ### Patch Changes