+++ packages/scripts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/ea-scripts
 
+## 2.4.29
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-factories@1.2.26
+
 ## 2.4.28
 
 ### Patch Changes