+++ packages/sources/layer2-sequencer-health/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/layer2-sequencer-health-adapter
 
+## 2.3.7
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 2.3.6
 
 ### Patch Changes