+++ packages/sources/solana-view-function/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/solana-view-function-adapter
 
+## 2.2.27
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 2.2.26
 
 ### Patch Changes